// client/src/app/store/store.model.service.ts
import { Injectable } from '@angular/core';
import { Brand } from '../shared/models/brand';
import { Product } from '../shared/models/product';
import { StoreData } from '../shared/models/storeData';
import { Type } from '../shared/models/type';

@Injectable({
    providedIn: 'root',
})
export class StoreModelService implements StoreData {
  products: Product[] = [];
  brands: Brand[] = [];
  types: Type[] = [];
  selectedBrand: Brand | null = null;
  selectedType: Type | null = null;
  selectedSort = 'asc'; // default val
  search = '';
  currentPage = 1;
  page?: number;
  pageable: any;
  totalElements: number = 70;
  pageSize: number = 10;
}